Add PackageFetchException, ensure package is unambiguous
Raise a detailed exception instead of asserting that the fetch did not fail.

Fixes the TODO questioning what dpkg outputs for `dpkg-query -S *' when
multiple packages of the same architecture are included

Turns out, dpkg outputs the architecture when it is ambiguous
(i.e. Multi-Arch: same). Therefore, assert that the package we
are selecting was the only option.