Add fix_broken parameter to install_packages
The first stage might not need to fix broken packages, so
add an optional parameter to disable dpkg error fixing. Whether
or not it is going to be used is still to be determined.

Add some print statements telling the user what the script is doing
during the apt-mark stages of install.