Add license and TODO, fix changelog and control
setup.cfg is needed so the package can be built twice in a row with
pybuild and debuild. Otherwise, it throws an error saying the
SOURCES.txt and requires.txt were modified from source.

Update changelog to close the ITP bug (not received yet).

Use the phrase "a Debian install" instead of "your Debian install" in
the package long description.

Update README.md so the Python package long description has similar
contents to the Debian package.