Refactor utils and main, remove elf-arch usage
Move the main function in crossgrader to __main__.py and reference it in
setup.py.

Create utils submodule to contain all scripts previously named *_util.py.

Since elf-arch is not available on Debian Stretch, parse each ELF's
e_machine value and compare it to the value of /usr/bin/dpkg. This will
be a more accurate reflection of the initramfs' crossgrade status
as well, showing if any packages are in the original architecture even
if the original architecture will run on the target architecture (e.g.
i386 to amd64).